+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Redirect Test Changelog
 
+## Unreleased
+
+- Enhancements
+  - Added `-a, --auth` option to test domains behind basic authentication.
+
+- Bug fixes
+  - fixed queue not getting killed if a task has an error before the queue finishes.
+
 ## v1.0.0
 
 - ALL THE THINGS :beers: